move GraphMenu and GraphButton to shared
